# docs/api/swagger.yaml
openapi: 3.0.0
info:
  title: "Miami Movies: Auth"
  version: v1
servers:
  - url: http://127.0.0.1:8888
paths:
  /api/v1/user/registration:
    post:
      tags:
        - user
      summary: "Регистрация пользователя"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRegistration"
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserItem'
        '422':
          description: Unprocessable Content
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '500':
          description: Unexpected Server Error
  /api/v1/user/login:
    post:
      tags:
        - user
      summary: "Вход пользователя в аккаунт"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserAuth"
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserLogin'
        '500':
          description: Unexpected Server Error
        '403':
          description: Forbidden
        '422':
          description: Unprocessable Content
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
  /api/v1/user/logout:
    post:
      tags:
        - user
      summary: "Выход пользователя из аккаунта"
      security:
        - AccessToken: []
      responses:
        '200':
          description: Successful operation
  /api/v1/user/info:
    get:
      tags:
        - user
      summary: "Получение информации об авторизованном пользователе"
      security:
        - AccessToken: []
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserItem'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Unexpected Server Error
...
  /api/v1/user/password/change:
    post:
      tags:
        - user
      summary: "Изменение пароля пользователя"
      security:
        - AccessToken: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserPasswordChange"
      responses:
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserItem'
...
components:
  securitySchemes:
    AccessToken:
      type: apiKey
      in: header
      name: X-Access-Token
    RefreshToken:
      type: apiKey
      in: header
      name: X-Refresh-Token
  schemas:
    UserAuth:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
    TokenItem:
      type: object
      properties:
          access:
            type: string
          refresh:
            type: string
    UserLogin:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/UserItem'
        tokens:
          $ref: '#/components/schemas/TokenItem'
    UserRegistration:
      type: object
      properties:
        first_name:
          type: string
        last_name:
          type: string
        password:
          type: string
        email:
          type: string
      required:
        - username
        - password
        - email
    UserItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        first_name:
          type: string
        last_name:
          type: string
        email:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    UserPasswordChange:
      type: object
      properties:
        password:
          type: string
...
    ValidationError:
      type: object
      properties:
        error_fields:
          description: >
            Error codes:
             * `428` - Обязательный параметр отсутствует
             * `422` - Не корректное значение
             * `424` - Нарушение связанности
             * `409` - Объект с такими полями уже существует
          type: object
          properties:
            field1:
              type: integer
              enum: [428, 422, 424]
            field2:
              type: integer
              enum: [428, 422, 424]